EN-37610 collocation capacity overflow (#305)

+++ coordinator/src/test/scala/com/socrata/datacoordinator/secondary/config/mock.scala
@@ -2,7 +2,7 @@ package com.socrata.datacoordinator.secondary.config.mock
 
 import com.socrata.datacoordinator.secondary.config.{SecondaryGroupConfig => ISecondaryGroupConfig, StoreConfig => IStoreConfig}
 
-case class StoreConfig(storeCapacityMB: Int, acceptingNewDatasets: Boolean) extends IStoreConfig
+case class StoreConfig(storeCapacityMB: Long, acceptingNewDatasets: Boolean) extends IStoreConfig
 
 case class SecondaryGroupConfig(numReplicas: Int, instances: Map[String, StoreConfig]) extends ISecondaryGroupConfig
